pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "msbc-rbac"
version = "0.0.4"
description = "A comprehensive Role-Based Access Control (RBAC) system for Django applications with multi-tenant support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Krish Patel", email = "your.email@example.com"}
]
keywords = ["django", "rbac", "role-based-access-control", "permissions", "multi-tenant", "authorization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "Django>=5.0",
    "psycopg2-binary>=2.9.1",
    "djangorestframework>=3.12.4",
    "drf-spectacular>=0.21.0",
    "python-decouple==3.8",
]

[project.urls]
Homepage = "https://github.com/yourusername/django-rbac-core"
Documentation = "https://github.com/yourusername/django-rbac-core#readme"
Repository = "https://github.com/yourusername/django-rbac-core"
"Bug Tracker" = "https://github.com/yourusername/django-rbac-core/issues"

[tool.setuptools]
packages = ["core", "accounts"]
include-package-data = true

[tool.setuptools.package-data]
core = ["migrations/*.py", "templates/**/*", "static/**/*"]
accounts = ["migrations/*.py", "templates/**/*", "static/**/*"]
